+++ packages/node-core/CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Fix modulo filter not applying correctly with multiple dataSources (#2331)
 
 ## [8.0.0] - 2024-03-28
 ### Added